requirements.txt
tiktoken
revChatGPT
poe-api
py-cord
bardapi
requests
EdgeGPT